feat: add validate-url lambda draft (#329)
Move all the agent job into isolated AWS lambda so it can be run as callable serveless function.
This has 2 benefits:

Improve scaling capabilities by running an agent in an isolated ephemeral container
Reduce task-manager weight, that will simply become a scheduler to trigger these lambdas ; meaning that we should reduce cost of our infra.
The task-manager sends a SQS message to trigger the lambda, wait for its resolution and push back the result to the redis.
This doesn't impact the current flow, it simply add another layer to the execution.

⚠️ This PR ONY impacts the validator_url job, we'll do the others in a follow up.
It's still a nice way to verify that everything works as expected.

Related to linear.app/solight/issue/SOL-204/move-taskmanager-to-aws-lambda

Signed-off-by: Vasek - Tom C <tom.chauveau@epitech.eu>