ci: attempt to speed up PR's build process (#343)
Instead of having a unique run_id for the built image, we will try to
tag image by PR so we can reuse the built image accross the PR and
expect better performances.

By default, dockerbuildx creates an isolated builder that will not reuse cache layer,
so we change that by setting the driver to docker.


Signed-off-by: Vasek - Tom C <tom.chauveau@epitech.eu>